Add Tests and VillagerType# registerBiomeType
